Add files via upload
Made a program that allows users to create their own PIN codes for their bank cards. And each PIN code consists of digits in such a way that when the user enters a character, the program stops and outputs "Please enter a number" and when the user enters only digits, the program outputs "PIN code is created".